# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "REST API for Topcoder Submissions.\n\n### Pagination\n  Requests that\
    \ return multiple items will be paginated to 20 items by default. You can specify\n\
    \  further pages with the `page` parameter. You can also set a custom page\n \
    \ size up to 100 with the `perPage` parameter.\n\n  Pagination response data is\
    \ included in http headers. By Default, the response header contains links with\
    \ `next`, `last`, `first`, `prev` resource links.\n"
  version: "5.0.0"
  title: "Topcoder Submissions REST API"
  termsOfService: "http://topcoder.com/terms/"
  contact:
    url: "http://www.topcoder.com/support"
    email: "support@topcoder.com"
host: "api.topcoder.com"
basePath: "/v5"
...
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
security:
- Bearer: []
paths:
...
  /reviewTypes:
    get:
      tags:
      - "Review types"
      summary: "Get all review types."
      description: "Get all review types. Link headers are sent back and they have\
        \ rel set to prev, next, first, last and contain the relevant URL.\n\nResulted\
        \ collection of review types can be filtered using filter parameters `name`,\
        \ `isActive` (all filter parameters are optional and combined by the logical\
        \ operation `AND`).\n\n**Authorization:** Review types is accessible by roles\
        \ `admin` and `copilot`.\n"
      operationId: "getReviewTypes"
      parameters:
      - name: "page"
        in: "query"
        description: "The page number."
        required: false
        type: "integer"
        default: 1
      - name: "perPage"
        in: "query"
        description: "The number of items to list per page."
        required: false
        type: "integer"
        default: 20
        maximum: 100
      - name: "name"
        in: "query"
        description: "The name filter for review types."
        required: false
        type: "string"
      - name: "isActive"
        in: "query"
        description: "The active boolean flag filter for review types."
        required: false
        type: "boolean"
      - name: "If-None-Match"
        in: "header"
        description: "Optional. If used in the request, the If-None-Match request\
          \ header should utilize the ETag as previously supplied in the response.\n"
        required: false
        type: "string"
      - name: "If-Modified-Since"
        in: "header"
        description: "Optional. In the subsequent request, the If-Modified-Since request\
          \ header can be populated with the previously retrieved `Last-Modified`.\n"
        required: false
        type: "string"
      responses:
        200:
          description: "OK - the request was successful, the resource are returned."
          schema:
            type: "array"
            example:
            - id: "928a2ec51c69ef84f5dccc4f"
              name: "Iterative Review"
              isActive: true
            - id: "c36f666674ed76b0071ee322"
              name: "Spec Review"
              isActive: true
            - id: "583581ae0c02bad6f274939b"
              name: "Screening"
              isActive: true
            items:
              $ref: "#/definitions/ReviewType"
          headers:
            X-Next-Page:
              type: "integer"
              description: "The index of the next page"
            X-Page:
              type: "integer"
              description: "The index of the current page (starting at 1)"
            X-Per-Page:
              type: "integer"
              description: "The number of items to list per page"
            X-Prev-Page:
              type: "integer"
              description: "The index of the previous page"
            X-Total:
              type: "integer"
              description: "The total number of items"
            X-Total-Pages:
              type: "integer"
              description: "The total number of pages"
            ETag:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                Common methods of ETag generation include using a collision-resistant\
                \ hash function of the resource's content, a hash of the last modification\
                \ timestamp, or even just a revision number.\n\nIn the next request,\
                \ the If-None-Match request header should utilize the ETag as previously\
                \ supplied in the response.\n"
            Last-Modified:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                The Last-Modified response HTTP header contains the date and time\
                \ at which the origin server believes the resource was last modified.\
                \ It is used as a validator to determine if a resource received or\
                \ stored is the same. Less accurate than an ETag header, it is a fallback\
                \ mechanism.\n\nNormally a database or otherwise maintained `Last-Modified\
                \ date` is maintained for the resource, which is supplied in the Last-Modified\
                \ response header.\n\nIn the subsequent request, the If-Modified-Since\
                \ request header can be populated with the previously retrieved `Last-Modified`.\n"
            Link:
              type: "string"
              description: "Pagination link header."
        304:
          $ref: "#/definitions/NotModified"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "ReviewTypes"
    head:
      tags:
      - "Review types"
      summary: "Get only response status and headers information but no response body\
        \ for the endpoint.\n"
      description: "Get response status and headers information for the endpoint.\
        \ Link headers are sent back and they have rel set to prev, next, first, last\
        \ and contain the relevant URL.\n\nRequested review types can be filtered\
        \ using filter parameters `name`, `isActive` (all filter parameters are optional\
        \ and combined by the logical operation `AND`).\n\n**Authorization:** Review\
        \ types is accessible by roles `admin` and `copilot`.\n"
      operationId: "headReviewTypes"
      parameters:
      - name: "page"
        in: "query"
        description: "The page number."
        required: false
        type: "integer"
        default: 1
      - name: "perPage"
        in: "query"
        description: "The number of items to list per page."
        required: false
        type: "integer"
        default: 20
        maximum: 100
      - name: "name"
        in: "query"
        description: "The name filter for review types."
        required: false
        type: "string"
      - name: "isActive"
        in: "query"
        description: "The active boolean flag filter for review types."
        required: false
        type: "boolean"
      - name: "If-None-Match"
        in: "header"
        description: "Optional. If used in the request, the If-None-Match request\
          \ header should utilize the ETag as previously supplied in the response.\n"
        required: false
        type: "string"
      - name: "If-Modified-Since"
        in: "header"
        description: "Optional. In the subsequent request, the If-Modified-Since request\
          \ header can be populated with the previously retrieved `Last-Modified`.\n"
        required: false
        type: "string"
      responses:
        200:
          description: "OK - The request was successful."
          headers:
            X-Next-Page:
              type: "integer"
              description: "The index of the next page"
            X-Page:
              type: "integer"
              description: "The index of the current page (starting at 1)"
            X-Per-Page:
              type: "integer"
              description: "The number of items to list per page"
            X-Prev-Page:
              type: "integer"
              description: "The index of the previous page"
            X-Total:
              type: "integer"
              description: "The total number of items"
            X-Total-Pages:
              type: "integer"
              description: "The total number of pages"
            Link:
              type: "string"
              description: "Pagination link header."
            ETag:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                Common methods of ETag generation include using a collision-resistant\
                \ hash function of the resource's content, a hash of the last modification\
                \ timestamp, or even just a revision number.\n\nIn the next request,\
                \ the If-None-Match request header should utilize the ETag as previously\
                \ supplied in the response.\n"
            Last-Modified:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                The Last-Modified response HTTP header contains the date and time\
                \ at which the origin server believes the resource was last modified.\
                \ It is used as a validator to determine if a resource received or\
                \ stored is the same. Less accurate than an ETag header, it is a fallback\
                \ mechanism.\n\nNormally a database or otherwise maintained `Last-Modified\
                \ date` is maintained for the resource, which is supplied in the Last-Modified\
                \ response header.\n\nIn the subsequent request, the If-Modified-Since\
                \ request header can be populated with the previously retrieved `Last-Modified`.\n"
        304:
          $ref: "#/definitions/NotModifiedWithoutBody"
        400:
          $ref: "#/definitions/BadRequestWithoutBody"
        401:
          $ref: "#/definitions/UnauthorizedWithoutBody"
        403:
          $ref: "#/definitions/ForbiddenWithoutBody"
        500:
          $ref: "#/definitions/ServerErrorWithoutBody"
      x-swagger-router-controller: "ReviewTypes"
    post:
      tags:
      - "Review types"
      summary: "Create a review type."
      description: "Create a new review type.\n\n**Authorization:** Review type creation\
        \ is accessible by roles `admin` and `copilot`.\n"
      operationId: "createReviewType"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/ReviewType"
      responses:
        201:
          description: "Created - The request was successful and the resource is returned."
          schema:
            $ref: "#/definitions/ReviewType"
          headers:
            Location:
              type: "string"
              description: "The Location response header indicates the URL to redirect\
                \ a page to or created resource. It only provides a meaning when served\
                \ with a 3xx (redirection) or 201 (created) status response."
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        409:
          $ref: "#/definitions/Conflict"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "ReviewTypes"
  /reviewTypes/{reviewTypeId}:
    get:
      tags:
      - "Review types"
      summary: "Get the review type."
      description: "Get the review type by id.\n\n**Authorization:** Review type is\
        \ accessible by roles `admin` and `copilot`.\n"
      operationId: "getReviewType"
      parameters:
      - name: "reviewTypeId"
        in: "path"
        description: "review type id"
        required: true
        type: "string"
      - name: "If-None-Match"
        in: "header"
        description: "Optional. If used in the request, the If-None-Match request\
          \ header should utilize the ETag as previously supplied in the response.\n"
        required: false
        type: "string"
      - name: "If-Modified-Since"
        in: "header"
        description: "Optional. In the subsequent request, the If-Modified-Since request\
          \ header can be populated with the previously retrieved `Last-Modified`.\n"
        required: false
        type: "string"
      responses:
        200:
          description: "OK - the request was successful and the review is returned."
          schema:
            $ref: "#/definitions/ReviewType"
          headers:
            ETag:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                Common methods of ETag generation include using a collision-resistant\
                \ hash function of the resource's content, a hash of the last modification\
                \ timestamp, or even just a revision number.\n\nIn the next request,\
                \ the If-None-Match request header should utilize the ETag as previously\
                \ supplied in the response.\n"
            Last-Modified:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                The Last-Modified response HTTP header contains the date and time\
                \ at which the origin server believes the resource was last modified.\
                \ It is used as a validator to determine if a resource received or\
                \ stored is the same. Less accurate than an ETag header, it is a fallback\
                \ mechanism.\n\nNormally a database or otherwise maintained `Last-Modified\
                \ date` is maintained for the resource, which is supplied in the Last-Modified\
                \ response header.\n\nIn the subsequent request, the If-Modified-Since\
                \ request header can be populated with the previously retrieved `Last-Modified`.\n"
        304:
          $ref: "#/definitions/NotModified"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        404:
          $ref: "#/definitions/NotFound"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "ReviewTypes"
    head:
      tags:
      - "Review types"
      summary: "Get only response status and headers information but no response body\
        \ for the endpoint.\n"
      description: "Get response status and headers information for the endpoint.\
        \ It does not contain response body.\n\n**Authorization:** Review type is\
        \ accessible by roles `admin` and `copilot`.\n"
      operationId: "headReviewType"
      parameters:
      - name: "reviewTypeId"
        in: "path"
        description: "review type id"
        required: true
        type: "string"
      - name: "If-None-Match"
        in: "header"
        description: "Optional. If used in the request, the If-None-Match request\
          \ header should utilize the ETag as previously supplied in the response.\n"
        required: false
        type: "string"
      - name: "If-Modified-Since"
        in: "header"
        description: "Optional. In the subsequent request, the If-Modified-Since request\
          \ header can be populated with the previously retrieved `Last-Modified`.\n"
        required: false
        type: "string"
      responses:
        200:
          description: "OK - The request was successful."
          headers:
            ETag:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                Common methods of ETag generation include using a collision-resistant\
                \ hash function of the resource's content, a hash of the last modification\
                \ timestamp, or even just a revision number.\n\nIn the next request,\
                \ the If-None-Match request header should utilize the ETag as previously\
                \ supplied in the response.\n"
            Last-Modified:
              type: "string"
              description: "Optional and can be used according to HTTP specification.\n\
                The Last-Modified response HTTP header contains the date and time\
                \ at which the origin server believes the resource was last modified.\
                \ It is used as a validator to determine if a resource received or\
                \ stored is the same. Less accurate than an ETag header, it is a fallback\
                \ mechanism.\n\nNormally a database or otherwise maintained `Last-Modified\
                \ date` is maintained for the resource, which is supplied in the Last-Modified\
                \ response header.\n\nIn the subsequent request, the If-Modified-Since\
                \ request header can be populated with the previously retrieved `Last-Modified`.\n"
        304:
          $ref: "#/definitions/NotModifiedWithoutBody"
        400:
          $ref: "#/definitions/BadRequestWithoutBody"
        401:
          $ref: "#/definitions/UnauthorizedWithoutBody"
        403:
          $ref: "#/definitions/ForbiddenWithoutBody"
        500:
          $ref: "#/definitions/ServerErrorWithoutBody"
      x-swagger-router-controller: "ReviewTypes"
    put:
      tags:
      - "Review types"
      summary: "Update the review type."
      description: "Update the review type by id.\n\n**Authorization:** Review type\
        \ update is accessible only by `admin` role.\n"
      operationId: "updateReviewType"
      parameters:
      - name: "reviewTypeId"
        in: "path"
        description: "review type id"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/UpdatableReviewType"
      responses:
        200:
          description: "OK - The request was successful and the updated resource is\
            \ returned."
          schema:
            $ref: "#/definitions/ReviewType"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        404:
          $ref: "#/definitions/NotFound"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "ReviewTypes"
    delete:
      tags:
      - "Review types"
      summary: "Delete the review type."
      description: "Delete the review type.\n\n**Authorization:** Review type deletion\
        \ is accessible only by `admin` role.\n"
      operationId: "deleteReviewType"
      parameters:
      - name: "reviewTypeId"
        in: "path"
        description: "review type id"
        required: true
        type: "string"
      responses:
        204:
          $ref: "#/definitions/NoContent"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        404:
          $ref: "#/definitions/NotFound"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "ReviewTypes"
    patch:
      tags:
      - "Review types"
      summary: "Partially update the review type."
      description: "Allows to partially modify the review type with the provided request\
        \ body properties.\n\n**Authorization:** Partially modify of review type is\
        \ accessible only by `admin` role.\n"
      operationId: "partiallyUpdateReviewType"
      parameters:
      - name: "reviewTypeId"
        in: "path"
        description: "review type id"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/PartiallyReviewType"
      responses:
        200:
          description: "OK - The request was successful and the updated resource is\
            \ returned."
          schema:
            $ref: "#/definitions/ReviewType"
        400:
          $ref: "#/definitions/BadRequest"
        401:
          $ref: "#/definitions/Unauthorized"
        403:
          $ref: "#/definitions/Forbidden"
        404:
          $ref: "#/definitions/NotFound"
        500:
          $ref: "#/definitions/ServerError"
      x-swagger-router-controller: "ReviewTypes"
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  NoContent:
    description: "No Content - The server has successfully fulfilled the request and\
      \ that there is no additional content to send in the response payload body.\n"
  BadRequestWithoutBody:
    description: "Bad Request - Request validation error, unable to process the request\
      \ due to an error (e.g., malformed request syntax, size too large, invalid request\
      \ message framing, or deceptive request routing).\n"
  UnauthorizedWithoutBody:
    description: "Unauthorized - Occurs when the JWT token provided is either expired\
      \ or invalid.\n"
  ForbiddenWithoutBody:
    description: "Forbidden - The request is not allowed.\n"
  NotModifiedWithoutBody:
    description: "Not Modified - Indicates that the resource has not been modified\
      \ since the last request.\n"
  ServerErrorWithoutBody:
    description: "Server Error - While handling the request something went wrong server-side.\n"
...
  ReviewType:
    allOf:
    - type: "object"
      properties:
        id:
          type: "string"
          example: "928a2ec51c69ef84f5dccc4f"
          description: "The review type id."
          readOnly: true
    - $ref: "#/definitions/PartiallyReviewType"
    description: "The review type entity."
  PartiallyReviewType:
    type: "object"
    properties:
      name:
        type: "string"
        example: "Iterative Review"
        description: "The review type name."
      isActive:
        type: "boolean"
        example: true
        description: "The review type activity boolean flag."
    description: "The review type entity fields that can be partially updated."
    example:
      name: "Iterative Review"
      isActive: true
  UpdatableReviewType:
    allOf:
    - $ref: "#/definitions/PartiallyReviewType"
    description: "The review entity type fields that updates whole entity."
...
  Unauthorized:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Unable to authenticate the user."
        description: "The unauthorized error message."
    description: "The unauthorized error entity."
  NotFound:
    type: "object"
    properties:
      message:
        type: "string"
        example: "A resource with the name could not be found."
        description: "The not found error message."
    description: "The not found error entity."
  NotModified:
    type: "object"
    properties:
      message:
        type: "string"
        example: "There is no need to transfer a representation of the target resource\
          \ because the request indicates that the client, which made the request\
          \ conditional, already has a valid representation."
        description: "The not modified error message."
    description: "The not modified entity."
  ServerError:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Something went wrong while processing your request. We’re sorry\
          \ for the trouble. We’ve been notified of the error and will correct it\
          \ as soon as possible. Please try your request again in a moment."
        description: "The server error message."
    description: "The server error entity."
  BadRequest:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Invalid input."
        description: "The bad request error message."
    description: "The bad request error entity."
  Forbidden:
    type: "object"
    properties:
      message:
        type: "string"
        example: "You are not allowed to access the request."
        description: "The forbidden error message."
    description: "The permission error entity."
  Conflict:
    type: "object"
    required:
    - "message"
    properties:
      message:
        type: "string"
        example: "Creating a resource with a name already exists."
        description: "The conflict error message."
    description: "The conflict error entity."
...